small optimization fix v3

+++ src/AccountCreator.java
@@ -11,13 +11,13 @@ interface TestInterface2{
 
 public class AccountCreator {
 
-    public List<Account> getAccSet(long num){
+    public List<Account> getAccSet(int num){
         List<Account> list = new ArrayList<>();
 
         TestInterface1 l = ()-> {
             double rand = Math.random();
-            if(rand > 0.4) return (int) Math.round(Math.random() * 10000);
-            else if (rand > 0.3) return (int) Math.round(Math.random() * 1000000000);
+            if(rand > 0.4) return Math.toIntExact(Math.round(Math.random() * 10000));
+            else if (rand > 0.3) return Math.toIntExact(Math.round(Math.random() * 1000000000));
             else return 0;
         };
